+++ CHANGELOG.rst
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ==========
 
+* Enable edit PageContent.
+* Fix get Category in CategoryArticleList.get.
+* Fix template pagination for view search.
+* Fix request.is_ajax in views.
+
 4.0.4 (2026-04-14)
 ==================